Updated 3.7.1
New Avatar Swipe Movement - We have replaced the previous position avatar rotation (horizontal and vertical) with a swipe driven movement. If you swipe a distance less than 1/4th of your screen (horizontal or vertical) you will get a slower rotation. When you swipe more than a 1/4th of your screen, you will get about three times the rotation speed. This change is in the wtw_inputs.js mouse and touch controls (to set the swipe direction, swipe strength, and key pressed array for animations) and in the wtw_avatarfunctions.js swipeRotateAvatar function (to set the rotations).

Touch Timer Changes - Global variables for WTW.touchLeftTimer and WTW.touchRightTimer have been replaced by WTW.touchTimer. With the new avatar swipe movement, there is no longer a need to determine if the touch was left or right side of the screen.

Remove Global Mouse Move Variables - We removed the global WTW.mouseMoveX and WTW.mouseMoveY since adding the new avatar swipe movement, they is no longer needed. The new swipe uses WTW.mouseStartX and WTW.mouseStartY instead.

Avatar Key Runtime Effects Added - In the moveAvatar function and using the new starttime and endtime per animation event (key pressed), we now calculate how long a key is being held down. Using this zruntime variable, we can change the weights of an animation the longer the key is held. For example, when walking or running, the avatar would turn slowly making it easy to adjust and fine tune the direction, but after a second or two, the turn rate will increase to allow the avatar to fully rotate faster.

Added Start Time and End Time to Avatar Running Animations - We added a starttime and endtime properties to the zrunninganims[zanim] array. This allows us to set a time the animation began and ended. When it begins, endtime gets set to null. When an animation ends, the starttime is set to null. These settings are related to when an animation is active = 1, which means that the key, mouse or touch is currently being activated.

Running Events Notation - To assist in reducing code in key functions, we set a zrunningevents variable for zrunninganims[zevent] which effectively is zavatar.WTW.animations.running[zevent] in the moveavatar function.

Running Animations Notation - To assist in reducing code in key functions, we set a zrunninganims variable for zavatar.WTW.animations.running in the moveavatar function.

Added Color Validation to Babylon Files - In the add Babylon Files function, we added validation to check the colors and if it is not the correct format, it will use default settings.

Key in Colors on Molds - When keying in colors on molds, it would allow you to paste, but not type a new color. We added a set color vs a displayed color code so that if the displayed is not valid, it will still use the set color code default value.

Thing Analytics ID during Cancel - When you cancelled the Thing Information form, the analytics id value was being reset to an invalid name analytics instead of analyticsid causing an undefined result. This has been corrected.

Thing Analytics Reloading - When editing a 3D Thing Information, the Google Analytics was not reloading on the form. It has been added to the sequence and now works as expected.

New 3D Web Open Link - After you download a 3D Web in the browser, the link that opens the new 3D Web in the browser for edit is now taking you to the Information Page so that you can start by renaming the new 3D Web.

Share Process Corrected - When trying to share a 3D Web (Community, Building, Thing, or Avatar), the process was receiving a 500 server error. This was caused by the use of an assignment notation instead of equals. It has been corrected and is fully functional.

3D Model Details Format - When the page is adjusted, the 3D Models Details view information was overwriting its container. We adjusted the containers on the page to each use half of the available space.

Improved Avatar Movement - We improved the avatar movement (moveAvatar function) by changing the new maxweight variable to be flexible per type of movement. Starting with, while walking other functions are limited to .3 weight, running limits other animations to .2 weight, and jumping does not allow other animation weights. Effectively, while walking or running, if you also turn, it will still keep the walking or running animation as dominant while keeping the forward movement persistent.

Avatar Movement Variables - The avatar movement variables (zonwalk, zonrun, zonjump, etc) have been moved to a level outside and before the keys pressed iteration. This allows them to be set once per function, instead of possible multiple times within the keys pressed loop. This also includes the conditional checking and setting of each variable.

Set Shown Molds Processing - In the wtw_common.js file, setShownMolds function, the building molds were not being processed after the initial load of a scene. For example, when teleporting to another location in a 3D Community, the buildings would not reappear after they are removed the first time. We corrected the condition to process the building molds when it is in a 3D Community, following the initial load.

No Covering Molds - When a no covering mold (like 3D Models) hit the complete molds function, it would reset the check collisions variable to false (0). We changed the code to check for collisions equal 1 then check the covering before a chance to overwrite the setting. It is working as expected now.

CSG Collisions Settings - When creating a CSG (merge, subtract, or union), the resulting mold was set to check collisions. Now it gets set to the same check collisions setting as the main mold being modified by the second mold.

CSG Remove Check Collisions Setting - When using CSG, the code was defaulting to no collisions. This setting was removed.

3D Text Rendering Group - When adding 3D Text to a scene, the text was not showing in the right render group, causing it to be behind the other items in the scene. We set the renderinggroupid to 1 to correct the issue.

Sorted the Edit Action Zone List - When you select to edit an existing action zone in a scene, the list was not in any particular order. It is now sorted by name so that it is easier to find things on a long list.

Teleport Status and Conditions - In the checkActionZones function, the condition for the teleport was including the init cycle variable. This caused the teleport to only work during the init cycle of a new site visit. We removed the condition. In addition, the teleport to the same 3D Community was not resetting the status of being in the teleport action zone, so it would only work once. It is now being reset when the teleport function is executed. Note that this status does prevent the function from executing multiple times.