Updated 3.7.1
New Enter Login Button Menu - When you come to a new scene, we moved the Enter button to a new 3D Model menu called wtw-entermenu.babylon. We then added a button for Login so that you can choose to login before entering the scene (Enter selects a random avatar). We also removed the Enter button from the login menu 3D Model. This change effects the HUD Login code as well as the Login and Core files that call the menus.

Jump Collisions Correction - When jumping, there is an avatar base collision adjustment so that the avatar can jump up on something as they walk or run forward. When the avatar stood still and jumped, the correction was causing the avatar to sink into the ground then slowly correct itself by collision control. We disabled the correction during a jump vertically to rectify this issue.

3D Internet Menu Closing Corrected - When in Admin Mode and click on the 3D Internet Menu, when you select a sub menu item, then click the 3D Internet to close the menu group, it now closes the sub menu item properly. Before, only the first item page was being checked for closure.

Enable / Disable WalkTheWeb Broadcasts - Under the 3D Internet, Multiplayer and Chat settings, in Admin Mode, you can now Enable and Disable the WalkTheWeb Broadcasts.

Move Avatar handling Jumps - We changed the moveAvatar function on how it processes and identifies Jump commands. It now uses an array of possible Jump commands to set the animation weights.

Exclude Space Bar from Key Pressed Remove - The keyPressedRemove function is used to release keys from the active controls. It can be found in the wtw_inputs.js file. We added a condition to the code to not remove the space bar (keycode 32) so that pressing the space bar will continue the action to jump until the jump animation is complete. The code to remove the space bar was added to the end function of the animation. (when loading a jump type animation).

Load Animation End Cycle Function - When executing an animation loop, for example jump, we needed to add code to execute at the end of a cycle before it loops. This allows us to add code to clear the jump before it loops. THis same location can be used to add additional conditions at the end of a cycle before it loops. This can be found in the loadAvatarAnimations function in wtw_loadavatar.js.

Added Space Bar Click Jump - Previously, you had to hold down the space bar to jump and it was an endless loop. This meant that after a jump sequence, gravity would not take over until the space bar was released. Now, jump uses a click on the space bar and always ends at one cycle of the jump.

Removed Opposite Movement from Move Avatar Section - Since the opposite movement cancellation happens on the key stroke end now, we do not need to test for it and cancel on the move avatar code side. We removed the checks and alternate fixes. This effects the wtw_avatarfunctions.js file.

Cancel Opposite Avatar Movements - In the keyPressedAdd function, we now check for opposite avatar movements and cancel the end results. For example, if left turn and right turn are both pressed, the avatar will do nothing (stay idle). Previously, it would jitter between both movements.

Canceling Key Repeat - When moving your avatar, if you hold a key, it was repeating the same process to check and add the key code. We now check to see if it is a repeating key before checking, making it run faster with less processing work.

Swap Avatar Control Keys - We swapped the default movement for the A, D, Q, and E keys. A and left arrow turn left, D and right arrow turn right, Q strafes left, and E strafes right. Note that R and F will look up or down and space bar jumps.

Teleport and Spawn in Copy, Share, and Downloads - The fields teleportwebid, teleportwebtype, and spawnactionzoneid needed to be added to every Copy, Share, and Download related function. It includes but was not limited to the connect, functions, handlers, scripts, and 3D Internet Plugin. The Copy, Share, and Download will now process and preserve these teleport and spawn settings as expected.

Teleport and Spawn Settings when Sharing - When you share a 3D Community, any teleport and spawn settings were not being transferred and included. We now added the teleportwebid, teleportwebtype, and spawnactionzoneid as part of the shared action zones for the community. Note that when downloading a shared 3D Community, only teleports within the same 3D Community will be included. Any additional teleport destinations will be ignored. This may change in the future, but it requires a domain name destination to stay consistent and we do not fully support that yet.

Save Action Zone with Teleport - When saving an action zone not of type teleport, the teleport community was not being cleared from previous use of the form, therefore would sometimes have values when it no longer applies. We updated the code to only set a value for the teleportwebid, teleportwebtype, and spawnactionzoneid if it is a type teleportzone.

Prevent Unloading for Background Tabs - When the browser tab is not the focus, browsers like Google Chrome would automatically unload some of the canvas causing 3D Models to disappear in the 3D Scene. Sometimes the land the avatar is standing on would disappear. This new setting for the engine to renderEvenInBackground will help prevent this from happening by keeping the rendering going in the background. Note that some browsers may throttle down the frames per second for background tabs.

Edit Water Render Corrected - In Admin Mode and working on a 3D Community (under Edit 3D Community, Edit 3D Scene, and Water, Waves, and Wind, select the Advanced Options), when editing the water, it was not rending the changes but removing the water. We now applied the changes using the same render code as the initial load, but with the new possible settings. This allows you to see the changes as you make them. Note that things like land, piers and other 3D objects will not show the refraction in the water until it is refreshed, they have to load after the water to be included in the refraction and it would take too long on each minute change to refresh everything.

Walk Distance and Animation Defaults - We changed the walk distance and animation defaults to 2.6 on the adjustment slider in the browse menu settings under avatar movement as well as the table default.

Walk Distance and Animation Range Increased - When adjusting the Avatar movement speeds in the browse menu settings, we increased the available range of values from .1 to 1 on the low end and 2.9 to 3.9 on the top end. We also increased the default setting from 1.5 to 2.6. This allows higher values for the avatar movement.

Avatar Distance Traveled Increased - We increased the factor that provides the distance traveled for avatars so that the walk, run, and strafe movements all move faster across the ground.