Updated 3.7.2
User Menu Buttons - The button naming for the Select Avatar and Edit Avatar buttons on the user menu were having a name conflict. We updated the 3D Menu Form and the HUD Login code to match the naming so it will work.

Check object before Setting Status - In the checkShownMolds function, the zloadazstatus is now checking to see if the object is null before trying to read the status.

Check Mold before Setting Collisions - We needed to add a check to see if the mold is null before setting the check collisions.

Avatar Animations Correction Globally - When the avatar has multiple animations for core motions such as walk, the avatar will have problems running and loading animations. This change to the connect file for avatar.php groups the animations by avatarevent for the ones that are not Options, and adds the Options grouped by friendly name, both only getting the latest updatedate when it finds duplicates. We corrected the same issue in Copy Avatar, Get Avatar Animations, Share and Download Avatars.

Avatar Animations Removing Duplicates - When the avatar has multiple animations for core motions such as walk, the avatar will have problems running and loading animations. This change to the connect file for avatar.php groups the animations by avatarevent for the ones that are not Options, and adds the Options grouped by friendly name, both only getting the latest updatedate when it finds duplicates.