Updated 3.7.2
Teleport to Other Communities - When teleporting to other communities, the teleport zone was firing multiple times, causing unexpected errors. This has been corrected by removing the status reset update step. Now it only happens once and gets the expected behavior.

Error Path Correction - In the class connecting grids functions, a number of functions were still referencing the previous class as the location for the error if it occurs. The path has been properly updated.

Communities Filter All - Reversed the last update, instead chose to use the filter all querystring to call the communities.php connect file. This allows the list of communities for Mine to only show ones I authored while giving all returned when supplying settings to forms like share and information.

Connect Communities for Admin - When opening communities.php from connect, if your role is admin and you open a community other than your own, the information was not necessarily showing. This update to the SQL fixes that issue and returns what it should.

Get User Roles - The function for getUserRoles was not returning more than one record in the array. It was missing the increment value. It has been corrected.

Download Queue and Update Badges - When you would Download or Cancel an item from the Download Queue (triggered from the downloads from www.walktheweb.org), the badge for downloads/updates was not being updated. This process now updates the badges.

First Building Cancel - When editing a 3D Community under Options and Settings, if you set the first building for the 3D Community, it removes the 3D Object from the scene. The Save and Cancel worked as planned, but the back arrow on the top was not. It is corrected to be treated as a cancel and remove the 3D Object.

Ser VR Browser User Agent - We used the following test for selecting the VR headset active: Oculus, Vibe, Mobile VR, and MobileVR.

Added VR Camera to XR Setup - We added the VR Camera with Gamepad to the initial XR loading environment. As a default, the VR Camera is now loaded when the browser user agent is VR.

XR Global Variables Added - We added 3 new global variables WTW.environment, WTW.xrHelper, and WTW.isVRorAR.

Enabled XR Environment - Added code to enable the XR Environment for VR and AR use. Currently, it only turns on if Oculus, Vibe, or Mobile VR is detected in the user agent.

Downloads for Teleporting - We decided to limit the Shared Downloads to only teleport within the same 3D Web. This code will set the teleportwebid to the correct new set id as needed when downloading.

Corrected Missing Image on Ground - The Extended Ground was missing an image because it only had an original image id and the SQL was checking for websizeid. It now returns an alternative image based on originalid. This update checks for the alternative and uses it if there is no path already defined.

Added Original Image to Response - Added the original image path to the recordset as an alternative when there is no websize id. This is applied to the extended ground texture, skybox texture, and water bump texture paths.

User Menu Buttons - The button naming for the Select Avatar and Edit Avatar buttons on the user menu were having a name conflict. We updated the 3D Menu Form and the HUD Login code to match the naming so it will work.

Check object before Setting Status - In the checkShownMolds function, the zloadazstatus is now checking to see if the object is null before trying to read the status.

Check Mold before Setting Collisions - We needed to add a check to see if the mold is null before setting the check collisions.

Avatar Animations Correction Globally - When the avatar has multiple animations for core motions such as walk, the avatar will have problems running and loading animations. This change to the connect file for avatar.php groups the animations by avatarevent for the ones that are not Options, and adds the Options grouped by friendly name, both only getting the latest updatedate when it finds duplicates. We corrected the same issue in Copy Avatar, Get Avatar Animations, Share and Download Avatars.

Avatar Animations Removing Duplicates - When the avatar has multiple animations for core motions such as walk, the avatar will have problems running and loading animations. This change to the connect file for avatar.php groups the animations by avatarevent for the ones that are not Options, and adds the Options grouped by friendly name, both only getting the latest updatedate when it finds duplicates.