Updated 3.8.2
Connect WordPress File Updated - When checking the wtw Access Token against the database for our WalkTheWeb WordPress plugin authentication, we reversed the logic to set the value as base 64 to compare to the raw database field value instead of the previous convert SQL command used in the database comparison. This has made it more reliable per versions of database used. In addition, some code was added to set a user as Host if it is being accessed on the 3d.walktheweb.com free initial hosting website.

Add User Role Added to WTWDB Class - The addUserRole function has been added to the WTWDB class functions. It also exists in the users class. It has also been extended to be used from the handlers and connect classes.

Role ID Is Valid Added to WTWDB Class - The roleIdIsValid function has been added to the WTWDB class functions. It also exists in the users class.

Get Role ID Added to WTWDB Class - The getRoleID function has been added to the WTWDB class functions. It also exists in the users class.

User ID Is Valid function Added to WTWDB - The userIdIsValid function has been added to the WTWDB class functions. It also exists in the users class.

Supports CTRL+U in Admin Mode - Added keycode 85 for U to the exception list for Admin Menu CTRL code shortcuts,. Now when you are in Admin Mode, you can use CTRL+U to open the page source.

Added Menu Form Support - The Key Down method in inputs was not processing the older style text fields. We added support to process them to send the keys. Some of the interpretations following the key addition will need to be modified to fully work as expected.

Shadow Quality Settings - When you use the settings, Shadow Quality slider to change the capabilities, the objects int he scene being updated were checking to see if they cast shadows before adjusting. Some objects do not use this setting and were causing an error in the update. Now the system checks for this setting availability before reading the setting only if it exists.

Added Easy Insert for 3D Models - When you are editing a 3D Community, Building, or Thing, we added a quick Add to Scene button on the 3D Models in the Library. This effectively is a shortcut to the other method of Admin Menu, Building, Edit Building, Add 3D Model, then selecting the model.

Corrected Physics fields references - In the copy community function, the references for the physics fields were not set properly, example, changed from $zphysicsenabled to $zrow["physicsenabled"] since the variable was not being set.

Water Level in Admin Mode Corrected - When in admin mode, the water level was not showing when you move the ground level down. This correction resolves that issue.

Get File From URL Updated - When downloading large files, like the WalkTheWeb Update, CURL handles the memory better than file_get_contents. Therefore, I reversed the order of preference of availability in the function.